Gentoro_interview/requirements.txt
mcp[cli]>=1.0.0
openai>=1.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
numpy>=1.24.0
python-dotenv>=1.0.0
